:white_check_mark: 테스트 코드와 API 문서에 캠페인 API&검색 API 통합 적용
